(fix) - O3-03130 - fix 'Add Provider Queue Room' dialog popping up when it should not
