+++ packages/esm-service-queues-app/src/patient-queue-metrics/clinic-metrics.test.tsx
@@ -67,8 +67,6 @@ describe('Clinic metrics', () => {
     expect(screen.getByText(/Average wait time today/i)).toBeInTheDocument();
     expect(screen.getByText(/minutes/i)).toBeInTheDocument();
     expect(screen.getByRole('button', { name: /queue screen/i })).toBeInTheDocument();
-    expect(screen.getByRole('button', { name: /add new service$/i })).toBeInTheDocument();
-    expect(screen.getByRole('button', { name: /add new service room/i })).toBeInTheDocument();
     expect(screen.getByText(/69/i)).toBeInTheDocument();
   });
 });
