+++ packages/esm-service-queues-app/translations/en.json
@@ -9,7 +9,7 @@
   "addNewService": "Add new service",
   "addNewServiceRoom": "Add new service room",
   "addPatientToQueue": "Add patient to queue",
-  "addProviderQueueRoom": "Add provider queue room?",
+  "addProviderQueueRoom": "Add provider queue room",
   "addQueue": "Add queue",
   "addQueueName": "Please add a queue name",
   "addQueueRoom": "Add queue room",
